Improve highstate returner
The highstate returner is able to "sprinkle" the generated mail,
which means that `{key}` is resolved at runtime to the grain `key`
on the minion rather than the master.

This is currently only done for the subject of the generated mail.

This change also calls sprinkle for the From and the To header, allowing
to use grains such as fqdn_host.